pyproject.toml
[project]
name = "esteban-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=1.24.0",
    "httpx>=0.27.0",
]

[project.scripts]
esteban = "agent:main"
